movies_admin/requirements.txt
django==3.2
djangorestframework==3.12.4
python-dotenv==1.0.0
psycopg2-binary==2.9.7
django-split-settings==1.2.0
gunicorn==20.1.0
